requirements.txt
tomli
pyChatGPT
openai
python-docx
requests